deps/v8/test/webkit/fast/js/Promise-static-cast-expected.txt

Test Promise.resolve as cast

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS promise is value
PASS result is undefined
PASS result2 is undefined
PASS result is "hello"
PASS result2 is 42
PASS fulfilled
PASS successfullyParsed is true

TEST COMPLETE
